api_revgas/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.4
PyMySQL==1.1.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1
